Add files via upload
Here's an update on what I have so far for my new project. I added a list of things at the end that I will continue to work on for the next few weeks. I'm having trouble with the first thing I mention there, as well as with the phylogenetic tree so maybe on Thursday if you have some time I could use some help. Thanks!